Add complete api server - Fix Bool_QA inference - Enhancing performance of first inference - Add demo shell script
